Extract HandleGeocoderSearch for use in other controllers
